+++ src/Samples/Providers/AzureOpenAI.cs
@@ -2,6 +2,7 @@
 using AgentFrameworkToolkit.OpenAI;
 using Azure.Identity;
 using Microsoft.Agents.AI;
+using Microsoft.Extensions.AI;
 using OpenAI.Chat;
 
 #pragma warning disable OPENAI001
@@ -13,11 +14,18 @@ public static class AzureOpenAI
     public static async Task RunAsync()
     {
         Configuration configuration = ConfigurationManager.GetConfiguration();
-        AzureOpenAIAgentFactory factory = new(new AzureOpenAIConnection
+        AzureOpenAIConnection connection = new()
         {
             Endpoint = configuration.AzureOpenAiEndpoint,
-            Credentials = new AzureCliCredential()
-        });
+            ApiKey = configuration.AzureOpenAiKey
+        };
+
+        AzureOpenAIEmbeddingFactory azureOpenAIEmbeddingFactory = new(connection);
+
+        IEmbeddingGenerator<string, Embedding<float>> embeddingGenerator = azureOpenAIEmbeddingFactory.GetEmbeddingGenerator("text-embedding-3-small");
+        GeneratedEmbeddings<Embedding<float>> generatedEmbeddings = await embeddingGenerator.GenerateAsync(["Hello"]);
+
+        AzureOpenAIAgentFactory factory = new(connection);
 
         AzureOpenAIAgent agent = factory.CreateAgent(new OpenAIAgentOptionsForChatClientWithReasoning
         {