+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 1.0.0-preview.251215.1
 - Added option to get Raw Client from the various Connection Objects (Except for the Google Connection as switch to the new official Google Nuget is expected soon)
 - OpenRouter and XAI Connection are now inherited from OpenAI Connections
+- OpenAI + AzureOpenAI: Added EmbeddingFactory
 
 ---